pokeapi/pyproject.toml
[tool.poetry]
name = "pokeapi"
version = "0.0.1"
description = "Poke-berries statistics API"
authors = ["William Méndez <wmendez.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
drf-spectacular = "^0.25.1"
gql = "^3.4.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
ipdb = "^0.13.11"
ipython = "^8.9.0"
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
werkzeug = "^2.2.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.6"

[tool.poetry.group.prod.dependencies]
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
